Don't use deprecated isc_interprete() in Firebird backend.
Using isc_interprete() is unsafe because it doesn't take the buffer size and
results in deprecation warnings with Firebird 2.x. Use the new fb_interpret()
instead to fix both problems.

Notice that this breaks compilation with Firebird 1.x but hopefully this is
not a problem as it's pretty old by now.

Signed-off-by: Vadim Zeitlin <vz-soci@zeitlins.org>