Use SQL_VARCHAR instead of SQL_LONGVARCHAR in ODBC backend.
This allows to use string parameters with equality operator when using MS SQL
ODBC driver. Before this change any attempt to compare a [VAR]CHAR(N) field in
the database with a string parameter resulted in the following error:

	ODBC error 42000: [Microsoft][ODBC SQL Server Driver][SQL Server]
	The data types char and text are incompatible in the equal to operator.

Also remove the strange SOCI_ODBC_VERSION_3_IS_TO_BE_CHECKED preprocessor
check which was useless as this symbol was never defined.

Signed-off-by: Vadim Zeitlin <vz-soci@zeitlins.org>