Remove dt_unsigned_long and x_unsigned_long type constants.
There is no need for a separate type for "unsigned long" as we don't have
special types neither for "unsigned int" nor for "long" and this type was not
handled correctly in the ODBC backend that mapped "unsigned int" to
SQL_BIGINT, even though the former is almost always 32 bits and the latter is
always 64 bits.

Simply remove these types to avoid the problem and simplify the code.

Signed-off-by: Vadim Zeitlin <vz-soci@zeitlins.org>