Add get_database_product() private method to ODBC backend.
This method allows to determine what kind of database are we actually
connected to. It is private as it's not clear whether it's a good idea to
expose it (although it might, in keeping with SOCI philosophy of letting to
use the underlying database API directly if needed), but it is needed to
implement the sequence functions in the ODBC backend, so do add it.

Signed-off-by: Vadim Zeitlin <vz-soci@zeitlins.org>