Add support for ODBC_OPTION_DRIVER_COMPLETE connection parameter.
Allow passing arbitrary options via connection_parameters and add support for
an option allowing to specify the "driver completion" mode, i.e. the degree of
interactivity of the driver, in the ODBC backend.

By default now do ask the user if the DSN doesn't include the user name and/or
password for the connection but the old behaviour is still available if
ODBC_OPTION_DRIVER_COMPLETE option is set to SQL_DRIVER_NOPROMPT before
opening the session.