Fix connection parsing in SOCI Firebird backend to handle spaces in values.
Really parse the connection strings instead of just breaking them on spaces.
This allows to detect errors such as not using "key=value" syntax at all but
most importantly allows to use spaces inside the values, as can be necessary
for the "service" parameter value when using embedded Firebird as the file
names can contain spaces.

Signed-off-by: Vadim Zeitlin <vz-soci@zeitlins.org>