Fix int and long mismatch for vector in ODBC
This is attempt to investigate and solve problems with failing ODBC tests after Travis CI switched to Linux VMs from 32-bit to  64-bit.
Apparently, there is mismatch of C++ types vector<int> and vector<long> int with use() and into(). ODBC backend relies on C++ sizeof and casting, so mixing 4-byte int with 8-byte long on Linxu 64-bit is asking for troubles.

This commit may be incomplete and serves purpose of illustration of the problem, needs to be reviewed and improved if needed. As the common-tests.h have been changed, it is necessary to verify those changes against other backends as well.
(Problems of this sort prove that sorting out strong requirements regarding C++ integer type support is a must.)