Fix regression with statement re-execution when using Firebird backend.
Since the introduction of endOfRowSet_ flag in 4c3cb629, executing the same
statement again stopped working as the flag was never reset after fetching all
the results of the first execution.

Fix this by resetting it in firebird_statement_backend::execute().

Also add a test verifying that this does indeed work.