Use type big enough to store unsigned long in Firebird tests.
Using decimal(9,0) for the "ul" field of the test database translates into
using SQL_LONG type at Firebird API level which, in turn, translates into
"int" at C++ level. And casting int values to unsigned long doesn't work
correctly on LP64 architectures where the former is 32 bits while the latter
is 64 bits.

This resulted in test failures when reading values > INT_MAX from the
database, as could be seen in the unit tests before 4c3cb629 masked it by
casting to unsigned int first. This, however, was inconsistent with the other
data types and broke reading of doubles from the database and will be undone
soon.

So use another fix for the unit tests: just create the field using "bigint"
type (alias for decimal(18,0)) which translates into SQL_INT64 and doesn't
truncate the value when casting it. This is also consistent with the other
backends that mostly use decimal(20,0) for this field.