Fix reading negative doubles from the database with Firebird backend.
Mantissa of type SQL_LONG must be converted to "int", not "unsigned", when
reading from the database.

This corrects the changes of 4c3cb629 which did use int in other places where
SQL_LONG is handled but not here to compensate for a different bug with
handling of unsigned long which is fixed differently by the previous commit.

Also add unit a (Firebird-specific for now but it should be made common) test
for reading negative doubles.