Don't allow exceptions escape from postgresql_statement_backend dtor.
This fixes abnormal program termination if an error happened while executing a
prepared statement: it was then destroyed during stack unwinding and if
deallocating it failed as well, resulting in a call to terminate().

In practice, deallocation seems to always fail with SQL state 25P02 ("current
transaction is aborted, commands ignored until end of transaction block") when
it is executed inside a transaction.