Disable test of define_and_bind sequence for vector
Apparently, ODBC backend (or drivers) are unable to cope with the
statement::execute following sequence of
undefine_and_bind/define_and_bind calls, leading to test_issue154
failures.
This feature is not high priority, I decided to move on with part of the
test_issue154 disabled until we investigate it further.
See discussion in #154.