Firebird: fix harmless warning in 64 bit builds.
In 64 bit builds Firebird defines isc_stmt_handle as an integer type and not a
pointer, so comparing it with or assigning NULL to it results in gcc warnings.

Avoid them by using 0 instead of NULL, this works in both 32 bit (where
isc_stmt_handle is a pointer) and 64 bit builds.