Use locale-independent function for converting doubles to strings.
Similarly to the issue #238 but in the other direction, we must use
locale-independent functions to ensure that we send the strings in correct
format to the database even if the current locale doesn't use period as
decimal separator.

This fixes errors like

	invalid input syntax for type double precision: "3,1415926500000002086"

when executing the tests in e.g. French locale.