Fix PostgreSQL unit test to pass with PostgreSQL < 9.0.
The output format for bytea values in PostgreSQL was "escape" until version 9
in which it was changed to "hex", which is what the current tests expect.

Explicitly check "bytea_output" parameter to make the tests pass with both 8
and 9 (alternative would be to always set it to "escape" on connection, but it
seems better to test the default behaviour).