Define GCC_WARNING_SUPPRESS and GCC_WARNING_RESTORE macros.
Add a private header for compiler-specific things and use it to define
GCC_WARNING_{SUPPRESS.RESTORE} macros which will be used in the upcoming
commits to locally disable some of the g++ warnings.

As a side effect, also add SOCI_CONCAT and SOCI_STRINGIZE helper macros, in
their own header to facilitate their reuse in the future if needed.