Don't start implicit transaction too eagerly in Firebird backend.
Firebird backend used to start a transaction immediately after construction or
after committing or rolling back the previous transaction. This created
problems due to unexpectedly not seeing any changes to the database not
performed via the same connection object because of the transaction isolation
and so SOCI_FIREBIRD_NORESTARTTRANSACTION existed to allow changing this
behaviour.

However it is not necessary to create the transactions so eagerly in the first
place, we can wait with doing it until the transaction handle is really
needed, i.e. when a statement is about to be executed. This solves the
problems due to having the implicit transaction open all the time and removes
the need for SOCI_FIREBIRD_NORESTARTTRANSACTION, simplifying the code.