Compare floating point values exactly in tests whenever possible.
The unit tests weaseled out of comparing the floating point values exactly
without any really valid excuse, change this to do check for the exact round
trip when possible.

Unfortunately it is not possible in quite a few cases: first, we can never
verify that performing "insert into t(d) values(3.14)" and then reading the
value back with "select d from t" returns the floating point constant 3.14f
because it is not equal to the string literal anyhow. This could, and probably
should, be fixed by using prepared statements whenever we use floating point,
and this is done for the insertions in the common test3() by this commit.

Second, some backends are simply buggy right now and don't write correct
values to the database. So for now enable the exact comparisons only for
Firebird for which the test does pass (at least with Firebird 2.5.2.26540
under Linux/amd64).