Fix rounding error in Firebird NUMERIC/DECIMAL conversions.
The value was truncated instead of being rounded in to_isc() helper function,
resulting in, for example, "0.003958" being saved as "0.003957" in a
NUMERIC(7,6) column because the string "0.003958" is actually
"0.0039579999999999997129" when converted to string with 20 digits of
precision and, after multiplying by 1e6 scale, yielded 3957 in the usual
double precision.

Round the values now to prevent this from happening.