Add use_type_base::dump_value() for richer diagnostics.
This method will allow including the values of the query parameters in the
error messages in the upcoming commits.

Also move get_name(), which was only available in standard_use_type before,
down to use_type_base, to allow getting the name of any parameter.

Notice that dump_value() implementation is pretty minimal currently, notably
it doesn't give any useful information about vectors and objects (i.e. mapped
to soci::value) parameters. This should be enhanced in the future.