Configure Travis CI to fail the build a.s.a.p.
There is no need to wait until the end of all builds if one of them fails
before marking the build as globally failing. Also, if we had any builds
allowed to fail, this would allow to mark the build as successful without
waiting for their end (but we don't have any such builds currently).