Don't add DB2_INCLUDE_DIR to global include directories.
This breaks the build of ODBC backend and the tests using it as DB2 headers
directory contains sqlext.h file which is also used by ODBC.

And it was unnecessary anyhow because the generic soci_backend() macro already
takes care of adding $BACKEND_INCLUDE_DIR to the include directories just for
the $BACKEND target.

Also remove similar link_directories() line -- even though it doesn't seem to
result in any problems, it's still unnecessary.