Fix a clash between CATCH test cases defined in different files.
Put the common test cases in their own namespace to avoid a clash between the
tests defined in soci::tests (which is brought into scope via using
directive by all the backend-specific files) and the tests defined in the
backend test sources if they happen to start at the same line.

This is just a workaround and a proper solution would, of course, be to move
the tests out of the header at all but this will have to wait.