Make SOCI_NORETURN public and use it in declaration.
SOCI_NORETURN had to be moved to a public header to avoid MSVC error C2381
("'function' : redefinition; __declspec(noreturn) differs") which was given if
SOCI_NORETURN was specified on the function definition only and not its
declaration.

Use SOCI_NORETURN when both declaring and defining the function now (notice
that C++11 [[noreturn]] attribute also must be used on declarations).