Use parameter names from the query, if given, in error messages.
Improve the error messages by using the parameter names used in the query
itself, if any, when the query parameters were bound by position and not by
name.

This means that errors for a query like

            sql << "insert into soci_test(name, age) values (:name, :age)",
                   use(name), use(age));

will now contain "with :name=... :age=..." even though the parameter names
were not passed to use().