Remove asserts from the SOCI headers and code.
Asserts shouldn't be used in the library code as it must not abort the main
application, whatever happens. Their use is especially problematic in the
headers as it can easily result in ODR violations if the library and the
application using it are compiled with different values of NDEBUG.

Add real error handling for a couple of asserts which are not trivially
verified (e.g. there is no need for checking that a pointer is not null after
initializing it with new as an exception would have been thrown anyhow if new
failed) and remove the rest.

Closes #300.