Extract common part of MySQL and ODBC MySQL tests in a header.
This makes it possible to reuse various MySQL-specific workarounds for both
tests without duplicating the not always trivial code.

As a side effect, fixes ODBC MySQL test which didn't pass any more since the
addition of exact double equality checks, making it use the same has_fp_bug()
as the MySQL test skips the failing test for it too now.