Removed dummy and not compilable operator<<(boost::optional).
"inline namespace" doesn't compile in C++98 mode, so the old code broke
compilation with g++ 4.7 or later or clang when not using C++11.

It also seems to be completely useless as outputting "Currently not supported"
just can't be a good idea, so it is not clear why was it added in the first
place (see PR #336).