Use noexcept(false) for once_temp_type dtor with MSVS 2015 too
Ensure that noexcept(false) is used for the dtor that can throw exceptions
when using MSVS 2015, otherwise the program would simply abort when it
happened.

Also rename SOCI_ONCE_TEMP_TYPE_NOEXCEPT macro to a more generic
SOCI_NOEXCEPT_FALSE in case we need to reuse it elsewhere and define it in
soci-platform.h.