Fix libraries link order when linking tests statically
Link with the backend (static) library first, then with the core
library in order to ensure that the backend library can use functions
defined in the core library even if they are not used by the core
library itself: previously, such functions would be discarded by a
traditional Unix linker when processing the core library because they
wouldn't be referenced and then be missing, resulting in link errors,
when it processes the backend library.

This will allow sharing some code between the backends in the upcoming
commit.