Add CLOB support to Firebird backend
Reuse the existing BLOB support, however we now need to allocate a BLOB
inside firebird_standard_use_type_backend objects as, unlike when using
soci::blob objects, we don't have any way to associate backend-specific
data with the parameter otherwise.

This difference between CLOB and BLOB support is not ideal, but there
doesn't seem to be any way to avoid it.