Make XML comparison in the test work for SQL Server too
Adding a new line to work around the problem with Oracle didn't allow
this test to pass with SQL Server which strips the newline if it is
present.

So don't modify the input, but do strip the trailing newline from output
if it's present in it to accommodate Oracle, but to allow SQL Server to
work too.