Enable XML/CLOB tests for PostgreSQL via ODBC
Try to address the problem with PostgreSQL ODBC driver truncating all
"unknown" fields, such as XML ones, to 256 bytes by default by changing
the corresponding connection attribute if possible.

Also document the need to do it when defining the DSN as, currently,
doing it during run-time will often fail as this option has only been
added in a rather recent version of the driver.