Fix buffer overflow if ODBC connection message box is cancelled
SQLDriverConnect() may return SQL_NO_DATA, which is not recognized as an
error by is_odbc_error(), if the connection string is incomplete and a
message box asking the user to enter the user name and/or password is
shown and cancelled.

In this case, we still used strLength bytes of outConnString array, but
the array remained initialized and, worse, strLength itself was
uninitialized too, meaning that we could read way beyond the end of the
array with possibly fatal consequences.

Fix this in the minimally intrusive way by just initializing strLength
to 0 to ensure that we never overflow the array. The next commit will
add better handling of SQL_NO_DATA.