Fix DDL creation of SQLite3 table with date/time column.
When user-defined C++ types are used, an exception of type std::bad_cast in type_conversion::from_base function is obtained when a std::tm type is used.

If the row type is declared as DATETIME, the error disappears.

I have only tested the case with SQLite3 backend. The version of SQLite was 3.22.0.