Fix using uninitialized sqlite3_column member
Parameters to SQLite can be provided in a sparse list (when not all
parameters are specified) .isNull_ field in sqlite3_column does not have
a default initializer. During the resize of a parameters vector isNull_
field for omitted parameters points to uninitialized memory, it is
interpreted as true most of the time, but not always. In release builds
this causes intermittent crashes or errors.

Closes #677.