Fix wrong SOCI_POSTGRESQL_NOSINGLEROWMODE check
Commit 09c26314dc9f7a8fe18249c4fedf67298d54c15d broke the use of single
row for PostgreSQL as any attempt to enable it at run-time would result
in an exception if it _was_ supported at compile-time (and if it wasn't
supported at compile-time, it couldn't work during run-time anyhow, of
course, so it effectively never worked at all).

Fix this by reversing the direction of the test.

The breakage was unfortunate, but also begs a question about how useful
this feature is if nobody noticed that it was broken in 3.5 years since
the problematic commit. And this, in turn, makes one wonder if it's
worth complicating the code with all these compile- and run-time checks
for a feature that nobody uses anyhow.

See #571.