Fall back on SQL_DRIVER_NOPROMPT when pooling ODBC connections
The application may enable (sometimes even inadvertently, simply due to
its use of another ODBC library which does it, e.g. pyODBC does this by
default) the use of pooled ODBC connections before trying to open a SOCI
session.

In this case, only the use of SQL_DRIVER_NOPROMPT is allowed by ODBC
driver manager and attempts to connect with any other "completion"
parameter value result in ODBC error "HY110". As the code using SOCI
may be completely oblivious to the use of pooled connections (e.g. SOCI
is used by another library in the same application and has no knowledge
whatsoever about the first library), it makes sense to try to do our
best in this situation and automatically try to fall back on connecting
with SQL_DRIVER_NOPROMPT in this case.

This commit implements this solution by adding a special check for this
case.

Note that these changes are best viewed ignoring whitespace.