Recognize more Oracle errors as corresponding to connection loss
Set error category to connection_error for ORA-03113 which occurs if the
server is not running any longer, ORA-03114 which happens in case of a
network problem and also for ORA-12541 which seems similar to ORA-12162,
which was already mapped to connection_error.

This allow the manual "Reconnect" test to pass with Oracle backend.