Initialize indicator to i_null even when there is no data at all
It is unexpected and error-prone to leave the indicator uninitialized if
a query doesn't retrieve any data, as it's too easy to forget to check
for got_data() and have the code working unpredictably depending on the
initial value of the uninitialized variable.

So always initialize the indicator value to i_null, this can't do any
harm, avoids bugs due to not checking got_data() and makes the code
simpler in the common case when absence of data is the same as having
NULL value in the database.

See #28.