Fix binding of vector parameters of custom types for ODBC too
This generalizes the changes of d41e0313 (Fix binding of parameters of
custom types in ODBC backend., 2012-05-09) which fixed the same problem
for scalar parameters to the vector ones (better late than never!).

As in the other commit, ODBC functions can't be called unless we have
the real data already, i.e. after converting the custom type to the data
in the ODBC format, so this needs to be done in pre_use(), called
significantly later than define_and_bind().

This commit deliberately tries to minimize the number of changes and
keeps the essential parts, but the code could and should be further
refactored now and this will be done in upcoming commits.