Fix binding of vector parameters of custom types for DB/2 too
This is very similar to 3c7b1e7a (Fix binding of vector parameters of
custom types for ODBC too, 2021-03-11) for the ODBC backend and just
does the same thing for DB/2.

As with ODBC, this already worked correctly for scalar values (after
backporting the fix from them from ODBC), and now also works for the
vector ones.