Delay unloading of dynamic backends until they're no longer used
Calling dynamic_backends::unload() or unload_all() could result in
crashes if there were soci::session objects still using the backend
being unloaded in the same program and this was difficult to avoid as
this could be done by two different libraries not knowing about each
other (but both using SOCI).

Make calling unload() safer by not actually unloading the backend
immediately if it's in use and doing it later, when its use count drops
to 0.