Split build and test stages of the CI builds
While this doesn't matter for Travis CI, it makes sense to separate them
for the other CI systems using more structured approach to the build
steps.

Add a few test_*.sh scripts for the backends that need to do something
special when running the tests and just use run_test function for all
the others.