Fix using MySQL with GitHub Actions
Start the service explicitly as this is not done by default.

Also explicitly use password with mysql in GitHub Actions environment,
both for the root and for the current user.

This wasn't necessary on Travis CI but is required in this environment,
so specify the root password explicitly and also create the MySQL user
(which is not hardcoded "travis" any longer) and record its password in
~/.my.cnf.