Use full paths to PostgreSQL ODBC driver in odbcinst.ini
This fixes the problem with psqlodbca.so not being found by unixODBC in
GitHub Actions environment.

It remains unclear why was this unnecessary on Travis CI and why it
doesn't work out of the box, but without this change we were always
getting

[unixODBC][Driver Manager]Can't open lib 'psqlodbca.so' : file not found (SQL state 01000)

when trying to connect to the database.