Launch PostgreSQL database ourselves under macOS
It is installed there but not started by default, so just launch it
ourselves.

Note that we must not use PGHOST etc variables under Mac, as the server
seems to be configured to only listen on the local socket by default, so
don't define them globally and do it only for Linux manually instead.