Replace using PostgreSQL docker container with native server
"services" can't be used with macOS runners and there is no way to limit
them to Linux runners only (i.e. "if:" is not supported here), so we'd
have to use 2 different jobs and duplicate all the other steps if we
wanted to continue using them.

Instead, just install PostgreSQL manually under Linux. This also means
that it isn't done for the builds that don't need it uselessly any more.