Implement support for CLOBs in Oracle vector backends
Allow using vectors of long strings and XML values with Oracle and
enable the tests for them when using Oracle.

The implementation is quite inefficient because it parallels the
existing code for the scalar (standard) backends. It seems like we ought
to avoid copying potentially large amounts of data when dealing with
CLOBs, but currently we don't do it at all.