Clarify comment and slightly restructure convert_to_base()
Try to explain better why we may need to call the custom conversion in
one case and why we must not do it in the other one.

Also test for these two cases in two separate "if"s to emphasize that
they check different things.

No real changes in behaviour.