ODBC: Fix inserting null values with prepared statements
Don't set indHolder_ itself to SQL_NULL_DATA in this case, this will
result in any subsequent uses of the parameter to also insert null data,
even if the indicator value changes and it is not i_null any longer.

Instead, pass another indicator, which is always SQL_NULL_DATA only for
this use of the parameter.

This fixes the recently added "NULL with optional" test with ODBC
backend.