Fix working with default Oracle NUMBER using rowset API
Using rowset API with the result of to_number(), which returned a NUMBER
with both precision and scale of 0 (indicating that it's the default
NUMBER type, i.e. NUMBER(38, 10)), previously resulted in 'ORA-01455:
converting column overflows integer datatype' error because we
incorrectly handled such numbers as integers rather than doubles that
they really are.

Closes #872.

See #874, #875.

Co-Authored-By: Vadim Zeitlin <vz-soci@zeitlins.org>