Fix fetching by rows with vector intos
Bind a single value of a vector for simple types which directly binded as pointer
to the first element of a vector in odbc_vector_into_type_backend::define_by_pos().
Which didn't worked when we fetching by a row.

See 9b4c5484faa4a753b6bdee19673f41912c5b5835