Use correct type for odbc_statement_backend::intos_ vector
For some reason this vector, added recently in 9b4c5484 (Fetch vectors
by a single row for huge strings in ODBC backend, 2021-03-24), used
"void*" for its elements instead of "odbc_vector_into_type_backend*",
requiring casts in several places.

Do use the actual type now, as there doesn't seem to be any problems
with doing this and it simplifies the code and makes it safer.