Get rid of vector<indicator> in odbc_vector_into_type_backend
This extra vector is not really necessary, the existing vector<SQLLEN>
used with SQLFetch() is enough and we just need to rebind the indicator
pointer to the correct element in it when fetch vectors by rows.

This is is simpler (it notably allows to get rid of the hack forcing the
index to 0 in get_sqllen_from_vector_at()) and more obviously correct,
but doesn't really change the behaviour.