Goodbye Travis
Unfortunately, and despite continued statements to the contrary, Travis
is not usable for open source projects any longer as this involves
constantly nagging Travis support to manually top up the project credit,
which is just infeasible, so we have no choice but to stop using it.

It was nice while it lasted, thanks for all the fish.