soci::row use union type data_holder instead of RTTI
Because it is dangerous to use RTTI when sharing C++ objects across binary boundaries, so we'd better to use hard code type instead of RTTI for soci::row. Besides, when we need to change from one database to another, it is difficult to use the correct type for soci::row::get(). So we use boost::numeric_cast<> to do auto type conversion when WITH_BOOST is 'ON'.