Avoid -Wpedantic given for some versions of mysql.h
This warning, given due to a trailing comma in mysql.h, breaks a CI
build under Ubuntu 20.04.4.

Just suppress it, as it's harmless and we can't do anything about it
anyhow.

Note that this suppression is conditional and is activated by either
SOCI_MYSQL_SOURCE being predefined (which is the case when building SOCI
itself) or private soci-compiler.h header being already included, which
is used when building the unit tests. This is due to the fact that we
can't include this header unconditionally, as it's private and depends
on another private header, so making it public is not completely
trivial, but arguably should still be done later.