Map dt_date to datetime in SQLite3's DDL API
As the datetime data type doesn't exist in SQLite3, the DDL
method create_column_type() mapped dt_date to an integer type. This was
possible because SQLite3 internally handles datetime values as text,
real or integer. When using the soci::row API however, we want to be
able to query datetime columns as std::tm instead of integers.
To do this, create_column_type() must assign dt_date to a type that
gets mapped back to dt_date in get_data_type_map() in
sqlite3/statement.cpp. The most generic one in this case is datetime.