Make SOCI_xxx and WITH_xxx both work as expected
Dependency options are extremely confusing (see #935) ever since the
changes of eddaacb1 (handle WITH_{backend} switch to turn on/off deps,
2013-02-01), see #38, which added the WITH_ options conflicting with the
existing SOCI_ ones.

It's really not clear what is the best thing to do here, as by now there
are probably people using both of these options, so try to do the
hopefully least unexpected thing and disable the given dependency if
either of the options is specified.

This fixes a problem our own CI scripts because they used SOCI_xxx=OFF
to disable the various backends, which didn't work at all.