Revamp Postgresql BLOB implementation
This commit refactors the existing implementation to factor out common
code into dedicated functions.

Furthermore, this commit adds support for creating new BLOBs via the
soci::blob interface by writing to a default-constructed soci::blob
object. The created BLOB will be deleted again, unless it is inserted
into a database, in which case the lifetime of the created BLOB gets
extended indefinitely (or rather to the point at which the database
decides to destroy it).

Thus, it is now possible to use the soci::blob type to insert new BLOBs
in a Postgresql table.

Note: Attempting to insert a default-constructed soci::blob into a
database will also create a new BLOB object at the point at which Soci
decides to really perform the insert (after checking, e.g. passed
indicators).